Move Poi Block component to poi/blocks
