+++ src/panel/poi/blocks/Block.jsx
