+++ src/panel/poi/blocks/Phone.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Block from 'src/components/ui/Block';
+import Block from 'src/panel/poi/blocks/Block';
 
 const PhoneBlock = ({ block }) =>
   <Block className="block-phone" icon="icon_phone" title={_('Phone')}>
